fix FragmentNavigator move FAB to fragment
