+++ app/src/main/java/com/qoiu/dailytaskmotivator/presentation/main/Navigator.kt
@@ -26,8 +26,6 @@ interface Navigator {
             return fragment
         }
 
-        override fun updateFragment() {
-            fragment.update()
-        }
+        override fun updateFragment() = fragment.update()
     }
 }
